# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI FE Exam API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "operationId": "7e370449b00151b6422003b99ee09b88",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Articles retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "operationId": "b2cea13eed1b008f51325f20208f6207",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "link",
                                    "date",
                                    "content",
                                    "status",
                                    "writer_id",
                                    "editor_id",
                                    "company_id"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Title"
                                    },
                                    "link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/article"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-01"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Content"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Published"
                                    },
                                    "writer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "editor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a specific article",
                "operationId": "64e4fccf59ff45a0babae2b1c80bff4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update a specific article",
                "operationId": "00c68486d6a226fd5e4211d3de317793",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "link",
                                    "date",
                                    "content",
                                    "status",
                                    "writer_id",
                                    "editor_id",
                                    "company_id"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Title"
                                    },
                                    "link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/article"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-01"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Content"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Published"
                                    },
                                    "writer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "editor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete a specific article",
                "operationId": "dbffca84d7628668e71cb5944173d11e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "description": "Article model",
                "required": [
                    "image",
                    "title",
                    "link",
                    "date",
                    "content",
                    "status",
                    "writer_id",
                    "editor_id",
                    "company_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "image": {
                        "type": "string",
                        "example": "image"
                    },
                    "title": {
                        "type": "string",
                        "example": "title"
                    },
                    "link": {
                        "type": "string",
                        "example": "link"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2021-09-01"
                    },
                    "content": {
                        "type": "string",
                        "example": "content"
                    },
                    "status": {
                        "type": "string",
                        "example": "status"
                    },
                    "writer_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "editor_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "company_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}